fix: automatically remove unnecessary indents according to the settings,
docs: change file extension
