+++ src/components/view/sources/sources.ts
@@ -1,0 +1,21 @@
+import './sources.css';
+
+class Sources {
+  draw(data) {
+    const fragment = document.createDocumentFragment();
+    const sourceItemTemp = document.querySelector('#sourceItemTemp');
+
+    data.forEach((item) => {
+      const sourceClone = sourceItemTemp.content.cloneNode(true);
+
+      sourceClone.querySelector('.source__item-name').textContent = item.name;
+      sourceClone.querySelector('.source__item').setAttribute('data-source-id', item.id);
+
+      fragment.append(sourceClone);
+    });
+
+    document.querySelector('.sources').append(fragment);
+  }
+}
+
+export default Sources;
